+++ src/test/java/test/pet/DeletePetTest.java
@@ -16,7 +16,7 @@ public class DeletePetTest {
 
     @Steps
     private PetEndpoint petEndpoint;
-    private Integer createdPetId;
+    private long createdPetId;
 
     @Before
     public void createPet() {
